Apply remaining changes

+++ TheosMMTempKomaru/Source/BasicHacks.mm
@@ -31,8 +31,8 @@ void* BasicHacks::HacksThread(void* arg) {
                 KomaruPatch::WriteMem<uint32_t>(targetAddr, kTargetCoinValue);
                 
                 // Verify the write
-                uint32_t verifyValue = static_cast<uint32_t>(KomaruPatch::ReadMem(targetAddr));
-                if (verifyValue == kTargetCoinValue) {
+                uintptr_t verifyValue = KomaruPatch::ReadMem(targetAddr);
+                if (verifyValue == static_cast<uintptr_t>(kTargetCoinValue)) {
                     gPatchApplied.store(true, std::memory_order_relaxed);
                     gCurrentStatus.store("Coins Patched!", std::memory_order_relaxed);
                 } else {
